moving functionality from `calc.js` to a helper file
